+++ corvis/src/tracker/state_vision.h
@@ -228,7 +228,7 @@ struct state_camera: state_branch<state_node*> {
     state_branch<state_vision_group *> groups;
     void update_feature_tracks(const rc_ImageData &image);
     void clear_features_and_groups();
-    int feature_count() const;
+    size_t feature_count() const;
     int process_features(mapper *map, spdlog::logger &log);
     void remove_group(state_vision_group *g, mapper *map);
 
@@ -252,7 +252,7 @@ public:
     ~state_vision();
     uint64_t group_counter = 0;
 
-    int feature_count() const;
+    size_t feature_count() const;
     void clear_features_and_groups();
     state_vision_feature *add_feature(const tracker::feature_track &track_, state_vision_group &group);
     state_vision_group *add_group(const rc_Sensor camera_id, mapper *map);
